Tools/expand_ai/requirements.txt
Pillow==10.3.0
requests==2.32.3
